requirements.txt
chromadb==0.4.22
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.15
langchain-groq>=0.1.0
langgraph==0.2.45
langsmith>=0.3.45
openai==1.54.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
streamlit==1.29.0
numpy==1.24.3
